Userland: Link directly against LibUnicodeData where needed
This is partially a revert of commits:
    3a8e1cc68999195a1d28ec7dd47f3028e4662560
    daf993d7d7f8339b75d76c9af7d813d84b0de78b

Rather than adding the prot_exec pledge requried to use dlopen(), we can
link directly against LibUnicodeData in applications that we know need
that library.

This might make the dlopen() dance a bit unnecessary. The same purpose
might now be fulfilled with weak symbols. That can be revisted next, but
for now, this at least removes the potential security risk of apps like
the Browser having prot_exec privileges.